+++ packages/sdks/src/blocks/columns/columns.lite.tsx
@@ -6,11 +6,14 @@ import type { SizeName } from '../../constants/device-sizes';
 import RenderInlinedStyles from '../../components/render-inlined-styles.lite';
 import { TARGET } from '../../constants/target.js';
 import BuilderContext from '../../context/builder.context.lite';
+import type { Dictionary } from '../../types/typescript';
 
 type Column = {
   blocks: BuilderBlock[];
   width?: number;
 };
+
+type CSSVal = string | number;
 
 type StackColumnsAt = 'tablet' | 'mobile' | 'never';
 
@@ -46,9 +49,9 @@ export default function Columns(props: ColumnProps) {
       stackedStyle,
       desktopStyle,
     }: {
-      stackedStyle: string;
-      desktopStyle: string;
-    }): string {
+      stackedStyle: CSSVal;
+      desktopStyle: CSSVal;
+    }): CSSVal {
       return state.stackAt === 'tablet' ? stackedStyle : desktopStyle;
     },
 
@@ -56,9 +59,9 @@ export default function Columns(props: ColumnProps) {
       stackedStyle,
       desktopStyle,
     }: {
-      stackedStyle: string;
-      desktopStyle: string;
-    }): string {
+      stackedStyle: CSSVal;
+      desktopStyle: CSSVal;
+    }): CSSVal {
       return state.stackAt === 'never' ? desktopStyle : stackedStyle;
     },
 
@@ -69,11 +72,11 @@ export default function Columns(props: ColumnProps) {
         ? 'column-reverse'
         : 'column',
 
-    get columnsCssVars(): { [key: string]: string } {
+    get columnsCssVars(): Dictionary<string> {
       if (TARGET === 'reactNative') {
         return {
           flexDirection: state.flexDir,
-        };
+        } as Dictionary<string>;
       }
 
       return {
@@ -82,33 +85,33 @@ export default function Columns(props: ColumnProps) {
           stackedStyle: state.flexDir,
           desktopStyle: 'row',
         }),
-      };
+      } as Dictionary<string>;
     },
 
-    columnCssVars(index: number): { [key: string]: string } {
-      const width = state.getColumnCssWidth(index);
-      const gutter = `${index === 0 ? 0 : state.gutterSize}px`;
+    columnCssVars(index: number): Dictionary<string> {
+      const gutter = index === 0 ? 0 : state.gutterSize;
 
       if (TARGET === 'reactNative') {
         return {
-          width,
-          marginLeft: props.stackColumnsAt === 'never' ? gutter : '0',
-        };
+          marginLeft: props.stackColumnsAt === 'never' ? gutter : 0,
+        } as any as Dictionary<string>;
       }
 
+      const width = state.getColumnCssWidth(index);
+      const gutterPixels = `${state.gutterSize}px`;
       const mobileWidth = '100%';
-      const mobileMarginLeft = '0';
+      const mobileMarginLeft = 0;
 
       return {
         width,
-        'margin-left': gutter,
+        'margin-left': gutterPixels,
         '--column-width-mobile': state.getMobileStyle({
           stackedStyle: mobileWidth,
           desktopStyle: width,
         }),
         '--column-margin-left-mobile': state.getMobileStyle({
           stackedStyle: mobileMarginLeft,
-          desktopStyle: gutter,
+          desktopStyle: gutterPixels,
         }),
         '--column-width-tablet': state.getTabletStyle({
           stackedStyle: mobileWidth,
@@ -116,9 +119,9 @@ export default function Columns(props: ColumnProps) {
         }),
         '--column-margin-left-tablet': state.getTabletStyle({
           stackedStyle: mobileMarginLeft,
-          desktopStyle: gutter,
+          desktopStyle: gutterPixels,
         }),
-      };
+      } as any as Dictionary<string>;
     },
 
     getWidthForBreakpointSize(size: SizeName) {