+++ packages/sdks/src/functions/get-content/index.ts
@@ -7,15 +7,22 @@ import type { GetContentOptions } from './types.js';
 export async function getContent(
   options: GetContentOptions
 ): Promise<BuilderContent | null> {
-  return (await getAllContent({ ...options, limit: 1 })).results[0] || null;
+  const allContent = await getAllContent({ ...options, limit: 1 });
+  if ('results' in allContent) {
+    return allContent?.results[0] || null;
+  }
+
+  return null;
 }
 
-/**
- * TO-DO: Handle error responses.
- */
-interface ContentResponse {
-  results: BuilderContent[];
-}
+type ContentResponse =
+  | {
+      results: BuilderContent[];
+    }
+  | {
+      status: number;
+      message: string;
+    };
 
 export async function getAllContent(
   options: GetContentOptions
@@ -25,15 +32,24 @@ export async function getAllContent(
   const res = await fetch(url.href);
   const content = await (res.json() as Promise<ContentResponse>);
 
+  if ('status' in content && !('results' in content)) {
+    console.error('[Builder.io]: Error fetching data. ', content, options);
+    return content;
+  }
+
   const canTrack = options.canTrack !== false;
-  if (
-    canTrack &&
-    // This makes sure we have a non-error response with the results array.
-    Array.isArray(content.results)
-  ) {
-    for (const item of content.results) {
-      await handleABTesting({ item, canTrack });
+  try {
+    if (
+      canTrack &&
+      // This makes sure we have a non-error response with the results array.
+      Array.isArray(content.results)
+    ) {
+      for (const item of content.results) {
+        await handleABTesting({ item, canTrack });
+      }
     }
+  } catch (e) {
+    console.error('[Builder.io]: Could not setup A/B testing. ', e);
   }
 
   return content;