+++ packages/sdks/output/react-native/CHANGELOG.md
@@ -1,3 +1,11 @@
+### 0.1.15
+
+- Fix: Removed Columns `width` style causing crash due to `calc()` CSS value.
+
+### 0.1.14
+
+- Fix: Columns styles causing crash due to `number` values being written as `string`s.
+
 ### 0.1.13
 
 - Fix: Columns block styling & `stackColumnsAt` prop