+++ packages/sdks/output/vue/CHANGELOG.md
@@ -1,3 +1,11 @@
+### 0.1.15
+
+- No changes.
+
+### 0.1.14
+
+- No changes.
+
 ### 0.1.13
 
 - Fix: Columns block styling & `stackColumnsAt` prop